Move group set find method to the group set service
