+++ tests/unit/lms/services/course_test.py
@@ -280,39 +280,6 @@ class TestCourseService:
             update_columns=["updated"],
         )
 
-    @pytest.mark.usefixtures("course_with_group_sets")
-    @pytest.mark.parametrize(
-        "params",
-        (
-            {"context_id": "context_id", "group_set_id": "ID", "name": "NAME"},
-            {"context_id": "context_id", "name": "NAME"},
-            {"context_id": "context_id", "name": "name"},
-            {"context_id": "context_id", "name": "NAME    "},
-            {"context_id": "context_id", "group_set_id": "ID"},
-        ),
-    )
-    def test_find_group_set(self, svc, params):
-        group_set = svc.find_group_set(**params)
-
-        assert group_set["id"] == "ID"
-        assert group_set["name"] == "NAME"
-
-    @pytest.mark.usefixtures("course_with_group_sets")
-    @pytest.mark.parametrize(
-        "params",
-        (
-            {"context_id": "context_id", "group_set_id": "NOID", "name": "NAME"},
-            {"context_id": "context_id", "group_set_id": "ID", "name": "NONAME"},
-            {"context_id": "no_context_id", "group_set_id": "ID", "name": "NAME"},
-        ),
-    )
-    def test_find_group_set_no_matches(self, svc, params):
-        assert not svc.find_group_set(**params)
-
-    @pytest.mark.usefixtures("course_with_group_sets")
-    def test_find_group_set_returns_first_result(self, svc):
-        assert svc.find_group_set()
-
     @pytest.mark.parametrize(
         "param,field",
         (
@@ -474,22 +441,6 @@ class TestCourseService:
             authority_provided_id=grouping_service.get_authority_provided_id.return_value,
             lms_id="context_id",
         )
-
-    @pytest.fixture
-    def course_with_group_sets(self, course):
-        course.extra = {
-            "group_sets": [
-                {
-                    "id": "ID",
-                    "name": "NAME",
-                },
-                {
-                    "id": "NOT MATCHING ID NOISE",
-                    "name": "NOT MATCHING NAME NOISE",
-                },
-            ]
-        }
-        return course
 
     @pytest.fixture
     def application_instance(self, application_instance):
