+++ tests/unit/lms/services/group_set_test.py
@@ -35,9 +35,68 @@ class TestGroupSetService:
             == group_set["name"]
         )
 
+    @pytest.mark.usefixtures("course_with_group_sets")
+    @pytest.mark.parametrize(
+        "params",
+        (
+            {"context_id": "context_id", "group_set_id": "ID", "name": "NAME"},
+            {"context_id": "context_id", "name": "NAME"},
+            {"context_id": "context_id", "name": "name"},
+            {"context_id": "context_id", "name": "NAME    "},
+            {"context_id": "context_id", "group_set_id": "ID"},
+        ),
+    )
+    def test_find_group_set(self, svc, params, application_instance):
+        group_set = svc.find_group_set(
+            application_instance=application_instance, **params
+        )
+
+        assert group_set["id"] == "ID"
+        assert group_set["name"] == "NAME"
+
+    @pytest.mark.usefixtures("course_with_group_sets")
+    @pytest.mark.parametrize(
+        "params",
+        (
+            {"context_id": "context_id", "group_set_id": "NOID", "name": "NAME"},
+            {"context_id": "context_id", "group_set_id": "ID", "name": "NONAME"},
+            {"context_id": "no_context_id", "group_set_id": "ID", "name": "NAME"},
+        ),
+    )
+    def test_find_group_set_no_matches(self, svc, params, application_instance):
+        assert not svc.find_group_set(
+            application_instance=application_instance, **params
+        )
+
+    @pytest.mark.usefixtures("course_with_group_sets")
+    def test_find_group_set_returns_first_result(self, svc, application_instance):
+        assert svc.find_group_set(application_instance)
+
     @pytest.fixture
     def svc(self, db_session):
         return GroupSetService(db=db_session)
+
+    @pytest.fixture
+    def course(self, application_instance):
+        return factories.Course(
+            application_instance=application_instance, lms_id="context_id"
+        )
+
+    @pytest.fixture
+    def course_with_group_sets(self, course):
+        course.extra = {
+            "group_sets": [
+                {
+                    "id": "ID",
+                    "name": "NAME",
+                },
+                {
+                    "id": "NOT MATCHING ID NOISE",
+                    "name": "NOT MATCHING NAME NOISE",
+                },
+            ]
+        }
+        return course
 
 
 class TestFactory:
